requirements.txt
fastapi
uvicorn[standard]
graphene
graphene-sqlalchemy
alembic